// protos/interfaces/auth/TokenExit.ts
// Original file: protos/auth.proto


export interface TokenExit {
  'email'?: (string);
  'name'?: (string);
  'role'?: (string);
  'status'?: (boolean);
  '_id'?: (string);
}

export interface TokenExit__Output {
  'email'?: (string);
  'name'?: (string);
  'role'?: (string);
  'status'?: (boolean);
  '_id'?: (string);
}