// protos/interfaces/auth/UserRegister.ts
// Original file: protos/auth.proto


export interface UserRegister {
  'email'?: (string);
  'password'?: (string);
  'name'?: (string);
  'traceContext'?: (string);
}

export interface UserRegister__Output {
  'email'?: (string);
  'password'?: (string);
  'name'?: (string);
  'traceContext'?: (string);
}