// protos/interfaces/auth/UserLoginOauth.ts
// Original file: protos/auth.proto


export interface UserLoginOauth {
  'email'?: (string);
  'password'?: (string);
  'name'?: (string);
  'photo'?: (string);
  'authType'?: (string);
}

export interface UserLoginOauth__Output {
  'email'?: (string);
  'password'?: (string);
  'name'?: (string);
  'photo'?: (string);
  'authType'?: (string);
}